Move each component to their own folders (to be used later). Remove leftover console.logs. Remove hisotry.js because it's unused.
