+++ massage-booking-system-frontend/src/components/AllAppointments/AllAppointments.js
@@ -1,18 +1,16 @@
 import React, { useContext } from 'react'
-import Appointment from './Appointment'
-import { AppointmentContext, UserContext } from '../App'
-
+import Appointment from '../Appoinment/Appointment'
+import { AppointmentContext, UserContext } from '../../App'
 
 const AllAppointments = () => {
   const { appointments, selectedDate, appointmentService } = useContext(AppointmentContext)
   const { users, user } = useContext(UserContext)
   const givenDate = new Date(selectedDate)
 
-
   let selectedDay = givenDate.getDate()
   let selectedMonth = givenDate.getMonth() + 1
   let selectedYear = givenDate.getFullYear()
-  // console.log('appointments in allappointments', appointments)
+
   // compares appointment time to selected date on calendar, filtering to only include selected days appointments
   const todaysAppointments = appointments.filter(appointment => {
     let appointmentsDate = new Date(appointment.start_date)
