+++ massage-booking-system-frontend/src/components/Appoinment/Appointment.js
@@ -1,7 +1,7 @@
 import React, { useContext } from 'react'
-import Display from './Display'
-import CreateAppointment from './CreateAppointment'
-import { AppointmentContext, UserContext, NotificationContext } from '../App'
+import Display from '../Display/Display'
+import CreateAppointment from '../CreateAppointment/CreateAppointment'
+import { AppointmentContext, UserContext, NotificationContext } from '../../App'
 
 const Appointment = props => {
   const { appointmentService } = useContext(AppointmentContext)
@@ -13,6 +13,7 @@ const Appointment = props => {
     await appointmentService.update(id, { type_of_reservation: 0, user_id: user._id, })
     createNotification('Appointment cancelled succesfully', 'success')
   }
+
   const markAppUnavailable = async () => {
     await appointmentService.update(id, '', 'remove')
   }
