+++ massage-booking-system-frontend/src/components/AppointmentList/AppointmentsLists.js
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react'
-import Appointment from './Appointment'
-import { AppointmentContext, UserContext } from '../App'
+import Appointment from '../Appoinment/Appointment'
+import { AppointmentContext, UserContext } from '../../App'
 
 const AppointmentsList = ({ ownPage }) => {
   const { appointments } = useContext(AppointmentContext)
