+++ massage-booking-system-frontend/src/components/NextAppointment/NextAppointment.js
@@ -1,12 +1,10 @@
 import React from 'react'
 
 const SimpleAppointment = ({ app }) => {
-
   const date = new Date(app.start_date)
   let minutes = date.getMinutes()
   let time = date.getTimezoneOffset()
   date.setMinutes(minutes + time)
-
 
   const weekdays = {
     0: 'Sunday',
