--- massage-booking-system-frontend/src/components/NotFoundPage.js
@@ -1,7 +1,0 @@
-import React from 'react'
-
-const NotFoundPage = () => {
-  return <div>404 - Page not found</div>
-}
-
-export default NotFoundPage
