+++ massage-booking-system-frontend/src/hooks/useResource.js
@@ -5,20 +5,17 @@ const useResource = baseUrl => {
   const [resources, setResources] = useState([])
 
   const getAll = async () => {
-    //console.log('GET ALL')
     const response = await axios.get(baseUrl)
     setResources(response.data)
   }
 
   const create = async data => {
-    console.log(data)
     const newResource = await axios.post(baseUrl, data)
     const updatedResources = resources.concat(newResource.data)
     setResources(updatedResources)
   }
 
   const update = async (id, data, type = '') => {
-    //console.log('UPDATE')
     const updatedResource = await axios.put(`${baseUrl}/${id}/${type}`, data)
     if (updatedResource.data.hasOwnProperty('_id')) {
       setResources(
@@ -30,7 +27,6 @@ const useResource = baseUrl => {
   }
 
   const updateExpectMany = async (id, type = '') => {
-    //console.log('UPDATE')
     const updatedResources = await axios.put(`${baseUrl}/${id}/${type}`)
     const data = updatedResources.data
 
@@ -43,14 +39,12 @@ const useResource = baseUrl => {
   }
 
   const remove = async id => {
-    //console.log('REMOVE')
     await axios.delete(`${baseUrl}/${id}`)
     const updatedResources = resources.filter(resource => resource._id !== id)
     setResources(updatedResources)
   }
 
   const getOne = async id => {
-    //console.log('GET ONE')
     const response = await axios.get(`${baseUrl}/${id}`)
     return response.data
   }
@@ -64,7 +58,6 @@ const useResource = baseUrl => {
   }
 
   const getInterval = async (start, end) => {
-    //console.log('GET INTERVAL')
     const response = await axios.get(`${baseUrl}/${start}/${end}`)
     setResources(response.data)
